requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
fastapi==0.115.6
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.4
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-core==0.3.83
langchain-google-genai==2.0.8
langchain-text-splitters==0.3.11
langgraph==0.2.74
langgraph-checkpoint==2.1.2
langgraph-sdk==0.1.74
langsmith==0.3.45
multidict==6.7.1
numpy==2.4.2
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pandas==2.2.3
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycoingecko==3.2.0
pycparser==3.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.41.3
ta==0.11.0
tenacity==9.1.4
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarl==1.22.0
zstandard==0.23.0
